% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Validate that a string looks like a URL}
\usage{
is_valid_url(url)
}
\arguments{
\item{url}{Character vector of length 1 with the URL}
}
\value{
Logical \code{TRUE} or \code{FALSE} as to whether the string passes the basic pattern.
}
\description{
Cursory check that the input starts with \verb{http://} or
\verb{https://} and contains at least one dot in the domain portion.
}
\keyword{internal}
