% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_escape}
\alias{html_escape}
\title{Minimal HTML escaping for IDE dialog messages}
\usage{
html_escape(x)
}
\arguments{
\item{x}{A character vector of length 1}
}
\value{
The same vector but with \code{&}, \code{<}, and \code{>} escaped.
}
\description{
Helper function to escape HTML in RStudio \code{\link[rstudioapi:showDialog]{rstudioapi::showDialog()}} dialog rendering.
}
\keyword{internal}
